REFACTOR moved light struct and function onto its own file
